[Refactoring] Extract logging code into new pod AlphaWalletLogger
