+++ AlphaWalletFoundation.podspec
@@ -38,8 +38,8 @@ Pod::Spec.new do |spec|
   spec.dependency 'TrustWalletCore', '2.6.34'
   spec.dependency 'EthereumABI'
   spec.dependency 'BlockiesSwift'
-  spec.dependency 'CocoaLumberjack', '3.7.0'
   spec.dependency 'PaperTrailLumberjack/Swift'
+  spec.dependency 'AlphaWalletLogger'
   spec.dependency 'AlphaWalletAddress'
   spec.dependency 'AlphaWalletCore'
   spec.dependency 'AlphaWalletGoBack'
