+++ AlphaWalletLogger.podspec
@@ -1,0 +1,27 @@
+#
+# Be sure to run `pod lib lint AlphaWalletLogger.podspec' to ensure this is a
+# valid spec before submitting.
+#
+# Any lines starting with a # are optional, but their use is encouraged
+# To learn more about a Podspec see https://guides.cocoapods.org/syntax/podspec.html
+#
+
+Pod::Spec.new do |s|
+  s.name             = 'AlphaWalletLogger'
+  s.version          = '1.0.0'
+  s.summary          = 'Logger'
+  s.description      = 'Logger'
+  s.homepage         = "https://github.com/AlphaWallet/alpha-wallet-ios/tree/master/modules/AlphaWalletLogger"
+  s.license          = { :type => 'MIT', :file => 'LICENSE' }
+  s.author             = { "Hwee-Boon Yar" => "hboon@motionobj.com" }
+  s.social_media_url   = "https://twitter.com/hboon"
+  s.ios.deployment_target = '13.0'
+  s.swift_version    = '4.2'
+  s.platform         = :ios, "13.0"
+  s.source           = { :git => 'git@github.com:AlphaWallet/alpha-wallet-ios.git', :tag => "#{s.version}" }
+  s.source_files     = 'modules/AlphaWalletLogger/AlphaWalletLogger/**/*.{h,m,swift}'
+  s.pod_target_xcconfig = { 'SWIFT_OPTIMIZATION_LEVEL' => '-Owholemodule' }
+
+  s.frameworks       = 'Foundation'
+  s.dependency 'CocoaLumberjack', '3.7.0'
+end
