+++ modules/AlphaWalletLogger/AlphaWalletLogger/AlphaWalletLogger.h
@@ -1,0 +1,18 @@
+//
+//  AlphaWalletLogger.h
+//  AlphaWalletLogger
+//
+//  Created by Vladyslav Shepitko on 11.02.2022.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for AlphaWalletLogger.
+FOUNDATION_EXPORT double AlphaWalletLoggerVersionNumber;
+
+//! Project version string for AlphaWalletLogger.
+FOUNDATION_EXPORT const unsigned char AlphaWalletLoggerVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <AlphaWalletLogger/PublicHeader.h>
+
+
